feat: Retrieve list of items from database and display in main table, remove backlog table
Implemented functionality to retrieve the list of items from the database and display them in the main table on the page. Additionally, the backlog table was removed since this information is now displayed in the main table.

Changes made include:
- Created a new function to retrieve the list of items from the database
- Modified the main table to display the data retrieved from the database
- Removed the backlog table from the page

This solves the problem of having to check two different tables to see the status of items and improves the user experience by having all relevant information in one place.

BREAKING CHANGE: The backlog table has been removed from the page, so this information is no longer displayed separately.

Rebase master