feat: Show item information according to selected row using JavaScript, retrieve image from database for display when product is selected
Implemented functionality to show item information dynamically when a row is selected in the table, using JavaScript to avoid page reloading and make the interaction more dynamic. Additionally, the item's image is retrieved from the database so that it's available to display when the product is selected.

Changes made include:
- Created a new JavaScript function to handle item selection and display information dynamically
- Modified the HTML and CSS to accommodate the new JavaScript function and display the item information
- Modified the server-side code to retrieve the item's image from the database and send it to the client for display

This improves the user experience by making the interaction with the table more dynamic and avoiding page reloading when selecting an item. It also provides a better presentation of the items by displaying the image when the product is selected.

BREAKING CHANGE: None.