feat: Show item image and description in sidebar when clicked, add main ingredients table
Implemented functionality to show the item's image and description in the sidebar when it's clicked, as well as a table of main ingredients for the item.

Changes made include:
- Created a new JavaScript function to handle item clicks and display information in the sidebar
- Modified the HTML and CSS to accommodate the new JavaScript function and display the item information
- Added a new table to display the main ingredients for the item, and populated it with data from the database
- Modified the server-side code to retrieve the item's description and image from the database and send it to the client for display

This improves the user experience by providing more detailed information about the item when it's clicked, including an image and description. It also provides a table of main ingredients for the item, making it easier for the user to see what's included in the dish.

BREAKING CHANGE: None.