feat: Add conditional to show "short" status for items with low inventory
Added a new conditional in the items planner view to display a "short" status for items with low inventory. The status is displayed based on the comparison between the available quantity and the quantity sold. If the available quantity is less than the quantity sold, the item is considered "short" and the status is displayed in the view.

Files modified:
- app/views/items_planner/index.html.erb

Changes made:
- Added new conditional to display "short" status for items with low inventory

This change improves the items planner by providing users with a clear indication of which items are running low on inventory. This helps to optimize production by ensuring that items are only produced when there is sufficient inventory to meet demand.

BREAKING CHANGE: None.