feat: Order items by status, display ingredients in sidebar, and style status labels
Changed the items planner view to order the items by status, giving priority to items with low inventory. The ingredients are now displayed in the sidebar, and the status is calculated based on the average quantity and the quantity that can be produced. The status labels have also been styled to make them more visible.

Files modified:
- app/assets/stylesheets/pages/_items-planner.scss
- app/controllers/items_planner_controller.rb
- app/javascript/planner.js
- app/models/item.rb
- app/views/items_planner/index.html.erb

Changes made:
- Updated the items planner controller to order the items by status
- Modified the JavaScript code to display the ingredients in the sidebar
- Added new calculation for item status based on average quantity and quantity that can be produced
- Styled the status labels to make them more visible

This change improves the items planner by providing users with a clearer view of which items are in low inventory and need to be prioritized. The display of ingredients in the sidebar makes it easier to access and manage recipe information, while the styled status labels make it easier to see the current status of each item.

BREAKING CHANGE: None.